Fix Feeder handshake bug (#17)
Fix 
```
A replica with the name: xxx is already active with the Feeder:null HANDSHAKE_ERROR: Error during the handshake between two nodes. xxxxxx
        at com.sleepycat.je.rep.stream.ReplicaFeederHandshake.negotiateProtocol(ReplicaFeederHandshake.java:198)
        at com.sleepycat.je.rep.stream.ReplicaFeederHandshake.execute(ReplicaFeederHandshake.java:250)
        at com.sleepycat.je.rep.impl.node.Replica.initReplicaLoop(Replica.java:710)
        at com.sleepycat.je.rep.impl.node.Replica.runReplicaLoopInternal(Replica.java:486)
        at com.sleepycat.je.rep.impl.node.Replica.runReplicaLoop(Replica.java:413)
        at com.sleepycat.je.rep.impl.node.RepNode.run(RepNode.java:1871)
```

If the leader is blocked for 30 seconds (heartbeat timeout), the follower will trigger an election, but the old leader will be re-elected with a high probability. In this case, the follower will close the old replica-feeder connection and open a new replica-feeder connection to the same leader. But if the old connection is still not deleted during the new link handshake, it will report a duplicate connection error.

The `dup.getChannel().isOpen()` is not enough to judge the availability of the connection, we should check whether there is heartbeat in the old connection.

Signed-off-by: gengjun-git <gengjun@starrocks.com>